refactor: restructure tests
