+++ tests/MCP/test_search.py
